Enemies and held humans look at player
Added the same aiming code from the player to the humans.

Currently they swing all the way around when the alien goes past them but this might become irrelevant when pathfinding is setup.

Ideally, the human would rotate their entire selves 180, on a tween, based on either "target is behind them" or "target is in front of them" states