Shoot mechanic refactored, works on Humans
Well it seems to work anyway.

Added a Projectile Handler that receives a signal from Human abductees if they are detected by the Player Proximity area.

The Projectile Handler now holds all info about weapons. It also checks if the human is allowed to shoot. It then gives the relevant info to a  "shoot point"  which is spawned on the gun barrel, and carries out an interval that will eventually be determined by the Handler's shooting_weapon dictionary. For now it's the same interval code as has been used for a while.

The copter is not setup to work for this yet, nor the drone.